+++ Assets/Tests/InputSystem/CoreTests_Actions_Rebinding.cs
@@ -1652,6 +1652,67 @@ internal partial class CoreTests
         Assert.That(action.ReadValue<Vector2>(), Is.EqualTo(new Vector2(0.6f * 4, 0.6f * 4)).Using(Vector2EqualityComparer.Instance));
     }
 
+    private static void AssertParameterValue(InputAction action, string name, float value, InputBinding bindingMask)
+    {
+        var x = action.GetParameterValue(name, bindingMask);
+        Assert.That(x.HasValue);
+        Assert.That(x.Value.ToSingle(), Is.EqualTo(value));
+    }
+
+    // ISXB-1721 - Parameter overrides do not work if a binding has the empty string assigned as its name.
+    [TestCase(null, null)]
+    [TestCase("", "")] // ISXB-1721 error case.
+    [TestCase("a", "b")]
+    [Category("Actions")]
+    public void Actions_CanApplySpecificParameterOverrides_UsingSpecificBindingMask(string firstBindingName, string secondBindingName)
+    {
+        var mouse = InputSystem.AddDevice<Mouse>();
+        var gamepad = InputSystem.AddDevice<Gamepad>();
+
+        var action = new InputAction(name: "Look", type: InputActionType.PassThrough);
+        var binding1 = action.AddBinding("<Mouse>/delta", groups: "Mouse", processors: "scaleVector2(x=0.25,y=0.25)");
+        var binding2 = action.AddBinding("<Gamepad>/leftStick", groups: "Gamepad", processors: "scaleVector2(x=4,y=4)");
+        if (firstBindingName != null)
+            binding1.WithName(firstBindingName);
+        if (secondBindingName != null)
+            binding2.WithName(secondBindingName);
+        action.Enable();
+
+        // Make sure parameters correspond to our binding definitions
+        AssertParameterValue(action, "x", 0.25f, action.bindings[0]);
+        AssertParameterValue(action, "y", 0.25f, action.bindings[0]);
+        AssertParameterValue(action, "x", 4.0f, action.bindings[1]);
+        AssertParameterValue(action, "y", 4.0f, action.bindings[1]);
+
+        // Override with the same mask as we fetched the parameter
+        action.ApplyParameterOverride("x", 1.2f, action.bindings[0]);
+        action.ApplyParameterOverride("y", 0.2f, action.bindings[0]);
+
+        // Check the parameter again
+        AssertParameterValue(action, "x", 1.2f, action.bindings[0]); // overridden
+        AssertParameterValue(action, "y", 0.2f, action.bindings[0]); // overridden
+        AssertParameterValue(action, "x", 4.0f, action.bindings[1]); // unmodified
+        AssertParameterValue(action, "y", 4.0f, action.bindings[1]); // unmodified
+
+        // Check that binding processor overrides are applied
+        Set(mouse.delta, new Vector2(-1.0f, -2.0f));
+        Assert.That(action.ReadValue<Vector2>(),
+            Is.EqualTo(new Vector2(-1.0f * 1.2f, -2.0f * 0.2f))
+                .Using(Vector2EqualityComparer.Instance));
+
+        // Check that binding processor overrides are not applied
+        Set(gamepad.leftStick, new Vector2(1.0f, 0.0f));
+        Assert.That(action.ReadValue<Vector2>(),
+            Is.EqualTo(new Vector2(1.0f * 4.0f, 0.0f * 4.0f))
+                .Using(Vector2EqualityComparer.Instance));
+
+        // Check that binding processor overrides are not applied
+        Set(gamepad.leftStick, new Vector2(0.0f, 1.0f));
+        Assert.That(action.ReadValue<Vector2>(),
+            Is.EqualTo(new Vector2(0.0f * 4.0f, 1.0f * 4.0f))
+                .Using(Vector2EqualityComparer.Instance));
+    }
+
     [Test]
     [Category("Actions")]
     public void Actions_CanApplyParameterOverrides_UsingBindingIndex()