+++ Packages/com.unity.inputsystem/Tests/TestFixture/InputTestFixture.cs
@@ -11,6 +11,8 @@ using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.InputSystem.Utilities;
 using UnityEngine.TestTools;
 using UnityEngine.TestTools.Utils;
+using UnityEngine.InputSystem.XR;
+
 #if UNITY_EDITOR
 using UnityEditor;
 using UnityEngine.InputSystem.Editor;
@@ -59,6 +61,17 @@ namespace UnityEngine.InputSystem
     /// input and device discovery or removal notifications from platform code. This ensures
     /// that while the test is running, input that may be generated on the machine running
     /// the test will not infer with it.
+    ///
+    /// Be cautious when using <c>NUnit.Framework.OneTimeSetUpAttribute</c> and
+    /// <c>NUnit.Framework.OneTimeTearDownAttribute</c> in combination with this test fixture.
+    /// For example, any devices created prior to execution of <see cref="Setup()"/> would be added to the actual
+    /// Input System instead of the test fixture system and after <see cref="Setup()"/> has executed such devices
+    /// will no longer be valid. You may of course use these NUnit features, but it is advised to not attempt affecting
+    /// the Input System under test from those methods since it would affect the real system and not the system
+    /// under test.
+    ///
+    /// This test fixture is designed for play-mode tests and is generally not supported for edit-mode tests.
+    /// Both <c>[Test]</c> and <c>[UnityTest]</c> are supported, but only in play-mode.
     /// </remarks>
     public class InputTestFixture
     {
@@ -533,6 +546,12 @@ namespace UnityEngine.InputSystem
         /// Note that this parameter will be ignored if the test is a <c>[UnityTest]</c>. Multi-frame
         /// playmode tests will automatically process input as part of the Unity player loop.</param>
         /// <typeparam name="TValue">Value type of the given control.</typeparam>
+        /// <exception cref="ArgumentNullException">If control is null.</exception>
+        /// <exception cref="ArgumentException">If the device associated with <paramref name="control"/> has not
+        /// been added to the system or if the control does not have any associated state. The latter may only
+        /// happen if attempting to set a control of a device created outside the test context.</exception>
+        /// <exception cref="NotSupportedException">If attempting to set a control of a test device in an
+        /// editor assembly. [UnityTest] in editor assemblies is not supported by this test fixture.</exception>
         /// <example>
         /// <code>
         /// var gamepad = InputSystem.AddDevice&lt;Gamepad&gt;();
@@ -544,12 +563,14 @@ namespace UnityEngine.InputSystem
         {
             if (control == null)
                 throw new ArgumentNullException(nameof(control));
-            if (!control.device.added)
-                throw new ArgumentException(
-                    $"Device of control '{control}' has not been added to the system", nameof(control));
+            CheckValidity(control.device, control);
 
             if (IsUnityTest())
+            {
+                if (IsEditMode())
+                    throw new NotSupportedException("InputTestFixture.Set does not support edit mode (editor assembly) [UnityTest].");
                 queueEventOnly = true;
+            }
 
             void SetUpAndQueueEvent(InputEventPtr eventPtr)
             {
@@ -663,6 +684,7 @@ namespace UnityEngine.InputSystem
                 if (screen == null)
                     screen = InputSystem.AddDevice<Touchscreen>();
             }
+            CheckValidity(screen);
 
             InputSystem.QueueStateEvent(screen, new TouchState
             {
@@ -708,32 +730,104 @@ namespace UnityEngine.InputSystem
                 throw new ArgumentException(
                     $"Action '{action}' must be bound to controls in order to be able to trigger it", nameof(action));
 
-            // See if we have a button we can trigger.
+            // We iterate controls to see if there are any that we know how to trigger
             for (var i = 0; i < controls.Count; ++i)
             {
-                if (!(controls[i] is ButtonControl button))
-                    continue;
+                var control = controls[i];
 
-                // Press and release button.
-                Set(button, 1);
-                Set(button, 0);
+                // for buttons, we literally trigger them pressed and released
+                if (control is ButtonControl buttonControl)
+                {
+                    Set(buttonControl, 1);
+                    Set(buttonControl, 0);
 
-                return;
+                    return;
+                }
+
+                // for the other types of controls we simply perform a small change in value as applicable
+                const float minorChange = 0.01f;
+
+                if (control is AxisControl axisControl)
+                {
+                    Set(axisControl, axisControl.ReadValue() + minorChange);
+
+                    return;
+                }
+
+                if (control is Vector2Control vec2Control)
+                {
+                    Set(vec2Control, vec2Control.ReadValue() + Vector2.one * minorChange);
+
+                    return;
+                }
+
+                if (control is Vector3Control vec3Control)
+                {
+                    Set(vec3Control, vec3Control.ReadValue() + Vector3.one * minorChange);
+
+                    return;
+                }
+
+                if (control is QuaternionControl quatControl)
+                {
+                    var q = quatControl.ReadValue();
+                    q.ToAngleAxis(out float angle, out Vector3 axis);
+                    Set(quatControl, Quaternion.AngleAxis(angle + minorChange, axis));
+
+                    return;
+                }
+
+                if (control is IntegerControl integerControl)
+                {
+                    Set(integerControl, integerControl.ReadValue() + 1);
+
+                    return;
+                }
+
+                if (control is TouchControl touchControl)
+                {
+                    var state = touchControl.ReadValue();
+                    state.position += Vector2.one * minorChange;
+                    Set(touchControl, state);
+
+                    return;
+                }
+
+                // this one is a bit weird, but there's not much to change other than picking the next phase in the list
+                if (control is TouchPhaseControl touchPhaseControl)
+                {
+                    var phase = touchPhaseControl.ReadValue();
+                    var values = Enum.GetValues(typeof(TouchPhase));
+                    var index = Array.IndexOf(values, phase);
+                    var newIndex = (index + 1) % values.Length;
+                    Set(touchPhaseControl, (TouchPhase)values.GetValue(newIndex));
+
+                    return;
+                }
+
+                if (control is BoneControl boneControl)
+                {
+                    var bone = boneControl.ReadValue();
+                    bone.position += minorChange * Vector3.one;
+                    Set(boneControl, bone);
+
+                    return;
+                }
+
+                if (control is EyesControl eyesControl)
+                {
+                    var eyes = eyesControl.ReadValue();
+                    eyes.leftEyePosition += minorChange * Vector3.one;
+                    Set(eyesControl, eyes);
+
+                    return;
+                }
             }
 
-            // See if we have an axis we can slide a bit.
-            for (var i = 0; i < controls.Count; ++i)
-            {
-                if (!(controls[i] is AxisControl axis))
-                    continue;
+            // Initially I wanted to implement PoseControl too, but it's got a very complicated ifdef that enables it.
+            // Didn't want to carry that #ifdef here. Let's see how many people really need to trigger PoseControl.
 
-                // We do, so nudge its value a bit.
-                Set(axis, axis.ReadValue() + 0.01f);
-
-                return;
-            }
-
-            ////TODO: support a wider range of controls
+            // If it's not a control that we know how to trigger - it's not implemented yet
             throw new NotImplementedException();
         }
 
@@ -912,6 +1006,39 @@ namespace UnityEngine.InputSystem
         }
 
         #endif
+
+        private static void CheckValidity(InputDevice device, InputControl control)
+        {
+            if (!device.added)
+            {
+                throw new ArgumentException(
+                    $"Device '{device}' has not been added to the system", nameof(device));
+            }
+
+            // Guards against a device from another scope being used. This is a direct way to evaluate whether
+            // the device is associated with the current manager state or not since device state isn't consistently
+            // pushed/popped in the current design.
+            var manager = InputSystem.s_Manager;
+            if (manager == null || !manager.HasDevice(device))
+            {
+                throw new ArgumentException($"Control '{control}' does not have any associated state. " +
+                    "Make sure the control or device was added after executing Setup().",  nameof(control));
+            }
+        }
+
+        private static void CheckValidity(InputControl control)
+        {
+            CheckValidity(control.device, control);
+        }
+
+        /// <summary>
+        /// Returns true if running inside an Edit Mode test (Editor assembly).
+        /// Returns false if running in Play Mode.
+        /// </summary>
+        private static bool IsEditMode()
+        {
+            return Application.isEditor && !Application.isPlaying;
+        }
 
         #if UNITY_EDITOR
         /// <summary>